[fix]  @JsonProperty(defaultValue="...") does not allow float/double values #152

+++ src/main/scala/com/kjetland/jackson/jsonSchema/JsonSchemaGenerator.scala
@@ -700,7 +700,7 @@ class JsonSchemaGenerator
           }
 
           extractDefaultValue(p).map { value =>
-            node.put("default", value.toInt)
+            node.put("default", value.toDouble)
           }
 
           // Look for @JsonSchemaExamples
